requirements.txt
absl-py==2.0.0
autopep8==2.0.4
azure-ai-formrecognizer==3.3.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
datasets
dill==0.3.7
filelock==3.13.1
fsspec==2023.10.0
future==0.18.3
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.2
huggingface-hub
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
Markdown==3.5.1
MarkupSafe==2.1.3
mpmath==1.3.0
multiprocess==0.70.15
munch==4.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
pdf2image==1.16.3
Pillow==10.1.0
protobuf==4.25.0
psutil==5.9.6
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.5.1
pydantic-settings==2.1.0
python-dateutil==2.8.2
python-doctr==0.7.0
pytorch-lightning==1.1.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
sacremoses==0.1.1
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
sconf==0.2.5
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
soupsieve==2.5
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
timm==0.5.4
tokenizers==0.14.1
tomli==2.0.1
torch==2.1.0
torch-tb-profiler==0.4.3
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3<2
weasyprint==60.1
Werkzeug==3.0.1
xxhash==3.4.1
zipp==3.17.0
zss==1.2.0
